refactor: 폴더명 변경 및 export 방식 named export로 변경
