+++ apps/community/src/components/professor-card.css.ts
